refactor: 서비스 테스트코드 성공, 실패로 분리하기 (#116)
